moved `eval_sh_jacobi` to `_poly.py`
